[Front] invitation-show + architecture app [Back] manifestation dates planification

+++ Sources/gestion-culturelle-back/src/main/java/com/afpa/cda/service/impl/ManifestationServiceImpl.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.afpa.cda.dao.AnimationRepository;
+import com.afpa.cda.dao.InvitationRepository;
 import com.afpa.cda.dao.ManifestationRepository;
+import com.afpa.cda.dao.ReservationRepository;
 import com.afpa.cda.dao.SalleRepository;
 import com.afpa.cda.dao.UserRepository;
 import com.afpa.cda.dto.AnimationDto;
@@ -18,7 +20,9 @@ import com.afpa.cda.dto.ManifestationDto;
 import com.afpa.cda.dto.SalleDto;
 import com.afpa.cda.dto.UserDto;
 import com.afpa.cda.entity.Animation;
+import com.afpa.cda.entity.Invitation;
 import com.afpa.cda.entity.Manifestation;
+import com.afpa.cda.entity.Reservation;
 import com.afpa.cda.entity.Salle;
 import com.afpa.cda.entity.User;
 import com.afpa.cda.service.IManifestationService;
@@ -37,6 +41,12 @@ public class ManifestationServiceImpl implements IManifestationService {
 
 	@Autowired
 	private UserRepository userRepository;
+	
+	@Autowired
+	private InvitationRepository invitationRepository;
+	
+	@Autowired
+	private ReservationRepository reservationRepository;
 
 	@Autowired
 	private ModelMapper modelMapper;
@@ -98,25 +108,25 @@ public class ManifestationServiceImpl implements IManifestationService {
 
 	@Override
 	public boolean findAvailability (ManifestationDto newManifestationDto) {
-
 		List<Manifestation> listManifestations =this.manifestationRepository.findAll();
+		long jourEnMs = 86400000;
 
 		for (Manifestation oldManifestation : listManifestations) {
+			System.out.println("oldManifestation.getId() "+oldManifestation.getId());
+			System.out.println("newManifestationDto.getDateDebut() "+newManifestationDto.getDateDebut().getTime()/jourEnMs);
+			System.out.println("newManifestationDto.getDateFin() "+newManifestationDto.getDateFin().getTime()/jourEnMs);
+			System.out.println("oldManifestation.getDateDebut() "+oldManifestation.getDateDebut().getTime()/jourEnMs);
+			System.out.println("oldManifestation.getDateFin() "+oldManifestation.getDateFin().getTime()/jourEnMs);
 			if (newManifestationDto.getId()!=oldManifestation.getId()) {
-				if (newManifestationDto.getSalle()!=null && newManifestationDto.getSalle().getLabel().equalsIgnoreCase(oldManifestation.getSalle().getLabel())) {
-					if ((newManifestationDto.getDateDebut().getTime()>=oldManifestation.getDateDebut().getTime() &&
-							newManifestationDto.getDateDebut().getTime()<=oldManifestation.getDateFin().getTime())
-							|| (newManifestationDto.getDateFin().getTime()>=oldManifestation.getDateDebut().getTime() &&
-							newManifestationDto.getDateFin().getTime()<=oldManifestation.getDateFin().getTime()))	{
-						System.out.println("newManifestationDto.getDateDebut() "+newManifestationDto.getDateDebut().getTime());
-						System.out.println("oldManifestation.getDateDebut() "+oldManifestation.getDateDebut().getTime());
-						System.out.println("newManifestationDto.getDateFin() "+newManifestationDto.getDateFin().getTime());
-						System.out.println("oldManifestation.getDateFin() "+oldManifestation.getDateFin().getTime());
-						
+				if (newManifestationDto.getSalle().getId()!=0 && newManifestationDto.getSalle().getId()==oldManifestation.getSalle().getId()) {
+					if (((newManifestationDto.getDateDebut().getTime()/jourEnMs-1)>=oldManifestation.getDateDebut().getTime()/jourEnMs &&
+							(newManifestationDto.getDateDebut().getTime()/jourEnMs-1)<=oldManifestation.getDateFin().getTime()/jourEnMs)
+							|| ((newManifestationDto.getDateFin().getTime()/jourEnMs-1)>=oldManifestation.getDateDebut().getTime()/jourEnMs &&
+							(newManifestationDto.getDateFin().getTime()/jourEnMs-1)<=oldManifestation.getDateFin().getTime()/jourEnMs))	{
 						return false;
 					}
 				}
-			} else return true;
+			} 
 		}
 		return true;
 	}
@@ -124,6 +134,11 @@ public class ManifestationServiceImpl implements IManifestationService {
 
 	@Override
 	public ManifestationDto findById(int id) {
+		try {
+			Thread.sleep(500);
+		} catch (InterruptedException e) {
+			e.printStackTrace();
+		}
 		Optional<Manifestation> manifOp = this.manifestationRepository.findById(id);
 
 		ManifestationDto manifestationDto = new ManifestationDto();
@@ -182,39 +197,39 @@ public class ManifestationServiceImpl implements IManifestationService {
 	public boolean add(ManifestationDto manifestationDto) {
 		Optional <Manifestation> manifestationOp = this.manifestationRepository.findManifestationByLabel(manifestationDto.getLabel());
 		if (!manifestationOp.isPresent()) {
-				
-		Manifestation manifestation = new Manifestation ();
-		manifestation.setLabel(manifestationDto.getLabel());
-		manifestation.setDateValidation(manifestationDto.getDateValidation());
 
-		Optional <User> validOp = this.userRepository.findById(manifestationDto.getValidateur().getId());
-		if (validOp.isPresent()) {
-			User valid = validOp.get();
-			manifestation.setValidateur(valid);
-		}
+			Manifestation manifestation = new Manifestation ();
+			manifestation.setLabel(manifestationDto.getLabel());
+			manifestation.setDateValidation(manifestationDto.getDateValidation());
 
-		Optional <Animation> animationOp = this.animationRepository.findById(manifestationDto.getAnimation().getId());
-		if (animationOp.isPresent()) {
-			Animation animation = animationOp.get();
-			manifestation.setAnimation(animation);
-		}
+			Optional <User> validOp = this.userRepository.findById(manifestationDto.getValidateur().getId());
+			if (validOp.isPresent()) {
+				User valid = validOp.get();
+				manifestation.setValidateur(valid);
+			}
 
-		manifestation.setDateDebut(manifestationDto.getDateDebut());		
-		manifestation.setDateFin(manifestationDto.getDateFin());
+			Optional <Animation> animationOp = this.animationRepository.findById(manifestationDto.getAnimation().getId());
+			if (animationOp.isPresent()) {
+				Animation animation = animationOp.get();
+				manifestation.setAnimation(animation);
+			}
 
-		Optional <Salle> salleOp = this.salleRepository.findById(manifestationDto.getSalle().getId());
-		if (salleOp.isPresent()) {
-			Salle salle = salleOp.get();
-			manifestation.setSalle(salle);
-		}
+			manifestation.setDateDebut(manifestationDto.getDateDebut());		
+			manifestation.setDateFin(manifestationDto.getDateFin());
 
-		manifestation.setReservations(manifestationDto.getReservations());
-		manifestation.setReservationsVip(manifestationDto.getReservationsVip());
-		manifestation.setCout(manifestationDto.getCout());
-		manifestation.setPrixBillet(manifestationDto.getPrixBillet());
+			Optional <Salle> salleOp = this.salleRepository.findById(manifestationDto.getSalle().getId());
+			if (salleOp.isPresent()) {
+				Salle salle = salleOp.get();
+				manifestation.setSalle(salle);
+			}
 
-		this.manifestationRepository.save(manifestation);
-		return false;
+			manifestation.setReservations(manifestationDto.getReservations());
+			manifestation.setReservationsVip(manifestationDto.getReservationsVip());
+			manifestation.setCout(manifestationDto.getCout());
+			manifestation.setPrixBillet(manifestationDto.getPrixBillet());
+
+			this.manifestationRepository.save(manifestation);
+			return false;
 		}
 		return true;
 	}
@@ -292,8 +307,10 @@ public class ManifestationServiceImpl implements IManifestationService {
 
 	@Override
 	public boolean delete(int id) {
-
-		if (this.manifestationRepository.existsById(id)) {
+		List <Invitation> listInvitations = this.invitationRepository.findInvitationByManifestation(id);
+		List <Reservation> listReservations = this.reservationRepository.findReservationByManifestation(id);
+		
+		if (listInvitations.isEmpty() && listReservations.isEmpty() && this.manifestationRepository.existsById(id)) {
 			this.manifestationRepository.deleteById(id);
 			System.err.println("manifestation supprimée");
 			return true;
